Adding password reset methods to the User model
Note that we’ve added a user_params method permitting both the password
and password confirmation attributes (Section 7.3.2: Strong parameters).

As noted above, the implementation in commit: The update action for
password reset delegates the boolean test for password reset expiration
to the User model via the code

```
@user.password_reset_expired?
```

To get this to work, we need to define the password_reset_expired?
method. As indicated in the email templates from Section 12.2.1:
Password reset mailer and templates, we’ll consider a password reset to
be expired if it was sent more than two hours ago, which we can express
in Ruby as follows:

```
reset_sent_at < 2.hours.ago
```

This can be confusing if you read < as “less than”, because then it
sounds like “Password reset sent less than two hours ago”, which is the
opposite of what we want. In this context, it’s better to read < as
“earlier than”, which gives something like “Password reset sent earlier
than two hours ago.” That is what we want, and it leads to the
password_reset_expired? method in commit: Adding password reset methods
to the User model. (For a formal demonstration that the comparison is
correct, see the proof in Section 12.6: Proof of expiration comparison)

With the code in this commit, the update action in commit: The update action for password reset should be working. (Lacking the patience to wait two hours, we’ll cover the third branch in a test, which is left as in up coming exercise not within this commit.)

1. Follow the email link from generated email and submit mismatched passwords to the form. What is the error message?

Started PATCH "/password_resets/R7DUwpOiYssK09gXhaRaDA" for 127.0.0.1 at 2017-12-30 00:19:                        46 -0500
Processing by PasswordResetsController#update as HTML
  Parameters: {"utf8"=>"✓", "authenticity_token"=>"iaZ5mzaAwAqrVv40hK6DLKdk3IUwztw7+hEqKQn                        +qLt7UF+Xvz3GPuKsnPpns+w64OqnQp9ILBxEBcWwKB44xg==", "email"=>"example-1@railstutorial.org"                        , "user"=>{"password"=>"[FILTERED]", "password_confirmation"=>"[FILTERED]"}, "commit"=>"Up                        date password", "id"=>"R7DUwpOiYssK09gXhaRaDA"}
  User Load (0.0ms)  SELECT  "users".* FROM "users" WHERE "users"."email" = ? LIMIT ?  [["                        email", "example-1@railstutorial.org"], ["LIMIT", 1]]
   (0.0ms)  begin transaction
  User Exists (0.0ms)  SELECT  1 AS one FROM "users" WHERE LOWER("users"."email") = LOWER(                        ?) AND ("users"."id" != ?) LIMIT ?  [["email", "example-1@railstutorial.org"], ["id", 2],                         ["LIMIT", 1]]
   (1.0ms)  rollback transaction
  Rendering password_resets/edit.html.erb within layouts/application
  Rendered shared/_error_messages.html.erb (1.0ms)
  Rendered password_resets/edit.html.erb within layouts/application (69.2ms)
  Rendered layouts/_rails_default.html.erb (286.8ms)
  Rendered layouts/_shim.html.erb (1.0ms)
  User Load (0.0ms)  SELECT  "users".* FROM "users" WHERE "users"."id" = ? LIMIT ?  [["id"                        , 2], ["LIMIT", 1]]
  Rendered layouts/_header.html.erb (3.0ms)
  Rendered layouts/_footer.html.erb (1.0ms)
Completed 200 OK in 1090ms (Views: 953.3ms | ActiveRecord: 1.0ms)

The form contains 1 error.

    Password is too short (minimum is 6 characters)

2. In the console, find the user belonging to the email link, and retrieve the value of the password_digest attribute. Now submit valid matching passwords to the form shown in Figure 12.12. Did the submission appear to work? How did it affect the value of password_digest? Hint: Use user.reload to retrieve the new value.

>> user = User.find_by_email('example-1@railstutorial.org')
user = User.find_by_email('example-1@railstutorial.org')
  User Load (0.0ms)  SELECT  "users".* FROM "users" WHERE "users"."email" = ? LIMIT ?  [["email", "example-1@railstutorial.org"], ["LIMIT", 1]]
=> #<User id: 2, name: "Deshawn Bins", email: "example-1@railstutorial.org", created_at: "2017-12-27 19:08:30", updated_at: "2017-12-30 05:19:32", password_digest: "$2a$10$B2JsqpKLRj8UBWlU9dVRVuCuUYblF9j780QBAIvuDR4...", remember_digest: nil, admin: false, activation_digest: "$2a$10$1UtcUeA/wW/5Hvi7An.lxOOGE81of29SO5cfDVZ31XE...", activated: true, activated_at: "2017-12-27 19:08:30", reset_digest: "$2a$10$GvpxYU4yW.qrbnPIDVBGv.jyJKoptmvjE2YUirORtUv...", reset_sent_at: "2017-12-30 05:19:32">
>> user.password_digest
user.password_digest
=> "$2a$10$B2JsqpKLRj8UBWlU9dVRVuCuUYblF9j780QBAIvuDR40quk4Ck1xa"
>> user.reload
user.reload
  User Load (1.0ms)  SELECT  "users".* FROM "users" WHERE "users"."id" = ? LIMIT ?  [["id", 2], ["LIMIT", 1]]
=> #<User id: 2, name: "Deshawn Bins", email: "example-1@railstutorial.org", created_at: "2017-12-27 19:08:30", updated_at: "2017-12-30 05:26:48", password_digest: "$2a$10$/38jlkopPOn0a5c2gvSZy.Z459ZoPkkpadHU0xGX9Gq...", remember_digest: nil, admin: false, activation_digest: "$2a$10$1UtcUeA/wW/5Hvi7An.lxOOGE81of29SO5cfDVZ31XE...", activated: true, activated_at: "2017-12-27 19:08:30", reset_digest: "$2a$10$GvpxYU4yW.qrbnPIDVBGv.jyJKoptmvjE2YUirORtUv...", reset_sent_at: "2017-12-30 05:19:32">
>> user.password_digest
user.password_digest
=> "$2a$10$/38jlkopPOn0a5c2gvSZy.Z459ZoPkkpadHU0xGX9GqTg9dBwpPK6"