An integration test for password resets.
The steps to test password resets broadly parallel the test for account activation from commit: Adding account activation to the user signup test, though there is a difference at the outset: we first visit the “forgot password” form and submit invalid and then valid email addresses, the latter of which creates a password reset token and sends the reset email. We then visit the link from the email and again submit invalid and valid information, verifying the correct behavior in each case. The resulting test, shown in this commit, is an excellent exercise in reading code.

Most of the ideas in this commit have appeared previously in this tutorial; the only really novel element is the test of the input tag:

```
assert_select "input[name=email][type=hidden][value=?]", user.email
```

This makes sure that there is an input tag with the right name, (hidden) type, and email address:

```
<input id="email" name="email" type="hidden" value="michael@example.com" />
```

With the code as in this commit, our test suite should be green:

$ rails test

Exercises -

1. In commit: Adding password reset methods to the User model, the create_reset_digest method makes two calls to update_attribute, each of which requires a separate database operation. By filling in the template shown in commit: A template for using update_columns, replace the two update_attribute calls with a single call to update_columns, which hits the database only once. After making the changes, verify that the test suite is still green. (For convenience, commit:  A template for using update_columns includes the results of solving the exercise in commit:  A template for using update_columns)

2. Write an integration test for the expired password reset branch in commit: The update action for password reset by filling in the template shown in commit: A test for an expired password reset. (This code introduces response.body, which returns the full HTML body of the page.) There are many ways to test for the result of an expiration, but the method suggested by commit: A test for an expired password reset is to (case-insensitively) check that the response body includes the word “expired”.

3. Expiring password resets after a couple of hours is a nice security precaution, but there is an even more secure solution for cases where a public computer is used. The reason is that the password reset link remains active for 2 hours and can be used even if logged out. If a user reset their password from a public machine, anyone could press the back button and change the password (and get logged in to the site). To fix this, add the code shown in commit: Clearing the reset digest on successful password reset to clear the reset digest on successful password update.

4. Add a line to commit: An integration test for password resets to test for the clearing of the reset digest in the previous exercise. Hint: Combine assert_nil (first seen in commit: A test of the “remember me” checkbox) with user.reload (commit: Adding account activation to the user signup test) to test the reset_digest attribute directly