The Micropost migration with added index.
In the course of developing the core sample application, we’ve now
encountered four resources—users, sessions, account activations, and
password resets—but only the first of these is backed by an Active
Record model with a table in the database. The time has finally come to
add a second such resource: user microposts, which are short messages
associated with a particular user.1 We first saw microposts in larval
form in Chapter 2, and in this chapter we will make a full-strength
version of the sketch from Section 2.3: The Microposts resource by
constructing the Micropost data model, associating it with the User
model using the has_many and belongs_to methods, and then making the
forms and partials needed to manipulate and display the results
(including, in Section 13.4, uploaded images). In the next branch:
following-users, we’ll complete our tiny Twitter clone by adding the
notion of following users in order to receive a feed of their
microposts.

A Micropost model

We begin the Microposts resource by creating a Micropost model, which
captures the essential characteristics of microposts. What follows
builds on the work from Section 2.3: The Microposts resource; as with
the model in that section, our new Micropost model will include data
validations and an association with the User model. Unlike that model,
the present Micropost model will be fully tested, and will also have a
default ordering and automatic destruction if its parent user is
destroyed.

If you’re using Git for version control, I suggest making a topic branch
at this time:

```
$ git checkout -b user-microposts
```

The Micropost model needs only two attributes: a content attribute to
hold the micropost’s content and a user_id to associate a micropost with
a particular user. The result is a Micropost model with the structure
shown in Figure 13.1.

It’s worth noting that the model in Figure 13.1 uses the text data type
for micropost content (instead of string), which is capable of storing
an arbitrary amount of text. Even though the content will be restricted
to fewer than 140 characters and hence would fit inside the
255-character string type, using text better expresses the nature of
microposts, which are more naturally thought of as blocks of text.
Indeed, we’ll use a text area instead of a text field for submitting
microposts. In addition, using text gives us greater flexibility should
we wish to increase the length limit at a future date (as part of
internationalization, for example). Finally, using the text type results
in no performance difference in production, so it costs us nothing to
use it here.

As with the case of the User model, we generate the Micropost model
using generate model with rails commands below:

```
$ rails generate model Micropost content:text user:references
```

This migration leads to the creation of the Micropost model.

refer to app/models/micropost.rb

In addition to inheriting from ApplicationRecord as usual, the generated
model includes a line indicating that a micropost belongs_to a user,
which is included as a result of the user:references argument in rails
code above. We’ll explore the implications of this line in Section
13.1.3: User/Micropost associations.

The generate command in Listing 13.1 also produces a migration to create
a microposts table in the database (in this commit); compare it to the
analogous migration for the users table from commit: Migration for the
User model (to create a users table). The biggest difference is the use
of references, which automatically adds a user_id column (along with an
index and a foreign key reference)3 for use in the user/micropost
association. As with the User model, the Micropost model migration
automatically includes the t.timestamps line, which (as mentioned in
Section 6.1.1: Database migrations) adds the magic created_at and
updated_at columns shown in Figure 13.1. (We’ll put the created_at
column to work starting in Section 13.1.4: Micropost refinements)

Because we expect to retrieve all the microposts associated with a given
user id in reverse order of creation, in this commit adds an index  on
the user_id and created_at columns:

```
add_index :microposts, [:user_id, :created_at]
```
By including both the user_id and created_at columns as an array, we
arrange for Rails to create a multiple key index, which means that
Active Record uses both keys at the same time.

With the migration in this commit, we can update the database as usual:

```
$ rails db:migrate
```
Exercises -

1. Using Micropost.new in the console, instantiate a new Micropost
object called micropost with content “Lorem ipsum” and user id equal to
the id of the first user in the database. What are the values of the
magic columns created_at and updated_at?

>> micropost = Micropost.new
micropost = Micropost.new
=> #<Micropost id: nil, content: nil, user_id: nil, created_at: nil,
updated_at: nil>
>> micropost.content = "Lorem ipsum"
micropost.content = "Lorem ipsum"
=> "Lorem ipsum"
>> micropost.user_id = User.first.id
micropost.user_id = User.first.id
  User Load (0.0ms)  SELECT  "users".* FROM "users" ORDER BY
"users"."id" ASC LIMIT ?  [["LIMIT", 1]]
=> 1

The magic column are both: nil

2. What is micropost.user for the micropost in the previous exercise?
What about micropost.user.name?

>> micropost.user
micropost.user
  User Load (1.0ms)  SELECT  "users".* FROM "users" WHERE "users"."id" =
? LIMIT ?  [["id", 1], ["LIMIT", 1]]
=> #<User id: 1, name: "Example User", email:
"example@railstutorial.org", created_at: "2017-12-27 19:08:30",
updated_at: "2017-12-27 19:08:30", password_digest:
"$2a$10$LaighSAvwXI451bqXpyWqemIThtPxHPAHHQcqnhpsN....",
remember_digest: nil, admin: true, activation_digest:
"$2a$10$.Qc4gexmoQwSw4torRVA0uZPUHL7LeBQgcAxMcI4MRw...", activated:
true, activated_at: "2017-12-27 19:08:19", reset_digest: nil,
reset_sent_at: nil>
>> micropost.user.name
micropost.user.name
=> "Example User"

3. Save the micropost to the database. What are the values of the magic
columns now?

>> micropost.save
micropost.save
   (0.0ms)  begin transaction
  User Load (0.0ms)  SELECT  "users".* FROM "users" WHERE "users"."id" =
? LIMIT ?  [["id", 1], ["LIMIT", 1]]
  SQL (1.0ms)  INSERT INTO "microposts" ("content", "user_id",
"created_at", "updated_at") VALUES (?, ?, ?, ?)  [["content", "Lorem
ipsum"], ["user_id", 1], ["created_at", "2017-12-30 14:35:14.802844"],
["updated_at", "2017-12-30 14:35:14.802844"]]
   (5.0ms)  commit transaction
=> true
>> micropost.created_at
micropost.created_at
=> Sat, 30 Dec 2017 14:35:14 UTC +00:00
>> micropost.updated_at
micropost.updated_at
=> Sat, 30 Dec 2017 14:35:14 UTC +00:00