Tests for the validity of a new micropost. green
Now that we’ve created the basic model, we’ll add some validations to
enforce the desired design constraints. One of the necessary aspects of
the Micropost model is the presence of a user id to indicate which user
made the micropost. The idiomatically correct way to do this is to use
Active Record associations, which we’ll implement in commit:
User/Micropost associations, but for now we’ll work with the Micropost
model directly.

The initial micropost tests parallel those for the User model (commit: A
test for validation of the name attribute). In the setup step, we create
a new micropost while associating it with a valid user from the
fixtures, and then check that the result is valid. Because every
micropost should have a user id, we’ll add a test for a user_id presence
validation. Putting these elements together yields the test in commit:
Tests for the validity of a new micropost.

As indicated by the comment in the setup method, the code to create the
micropost is not idiomatically correct, which we’ll fix in Section
13.1.3: User/Micropost associations.