A validation for the micropost’s user_id. green
As with the original User model test (commit: A test for an initially
valid user), the first test in commit: Tests for the validity of a new
micropost. is just a sanity check, but the second is a test of the
presence of the user id, for which we’ll add the presence validation
shown in this commit.

By the way, as of Rails 5 the tests in commit: Tests for the validity of
a new micropost actually pass without the validation in this commit, but
only when using the idiomatically incorrect line highlighted in commit:
Tests for the validity of a new micropost. The user id presence
validation is necessary after switching to the idiomatically correct
code in commit: Using idiomatically correct code to build a micropost,
so we include it here for convenience.

With the code in this commit the tests should (still) be green:

$ rails test:models

Started with run options --seed 44874

Progress: |Run options: --seed 44874

====.=====.=====..=========.=====.======.====.=====.=====.=====.====.=====.======|
.
Finished in 11.87629s
14 tests, 23 assertions, 0 failures, 0 errors, 0 skips

Finished in 12.037504s, 1.1630 runs/s, 1.9107 assertions/s.
14 runs, 23 assertions, 0 failures, 0 errors, 0 skips