A user has_many microposts. green
When constructing data models for web applications, it is essential to
be able to make associations between individual models. In the present
case, each micropost is associated with one user, and each user is
associated with (potentially) many microposts—a relationship seen
briefly in Section 2.3.3: A user has_many microposts and shown
schematically in Figure 13.2 and Figure 13.3. As part of implementing
these associations, we’ll write tests for the Micropost model and add a
couple of tests to the User model.

Figure 13.2: The belongs_to relationship between a micropost and its
associated user.

| id | content     | user_id |
-----------------------------|
| 3  | Lorem ipsum |    1    |-------------------|
                                                 |
                                                 |
  |--------------- belongs_to -------------------|
  |
  |
  |     |  id |   name        | email             |
  |     -------------------------------------------
  |-->  | 1  | Michael Hartl | mhart1@example.com |

Figure 13.3: The has_many relationship between a user and its
microposts.

| id | content     | user_id    |
--------------------------------|
| 3  | Lorem ipsum |    1       |<----has_many------|
                                                    |
                                                    |
| id | content        | user_id |                   |
----------------------| --------|                   |
| 4  | Dolor sit amet |    1    |<----has_many------|
                                                    |
                                                    |
| id | content     | user_id    |                   |
--------------------------------|                   |
| 7  | Consectetur |    1       |<----has_many------|
                                                    |
                                                    |
|---------------------------------------------------|
|
|
|     |  id |   name        | email             |
|     -------------------------------------------
|-----| 1  | Michael Hartl | mhart1@example.com |

Using the belongs_to/has_many association defined in this section, Rails
constructs the methods shown in Table 13.1. Note from Table 13.1 that
instead of:

```
Micropost.create
Micropost.create!
Micropost.new
```

we have

```
user.microposts.create
user.microposts.create!
user.microposts.build
```

These latter methods constitute the idiomatically correct way to make a
micropost, namely, through its association with a user. When a new
micropost is made in this way, its user_id is automatically set to the
right value. In particular, we can replace the code

```
@user = users(:michael)

@micropost = Micropost.new(content: "Lorem ipsum", user_id: @user.id)
```

```
@user = users(:michael)
@micropost = @user.microposts.build(content: "Lorem ipsum")
```

(As with new, build returns an object in memory but doesn’t modify the
database.) Once we define the proper associations, the resulting
@micropost variable will automatically have a user_id attribute equal to
its associated user’s id.

To get code like @user.microposts.build to work, we need to update the
User and Micropost models with code to associate them. The first of
these was included automatically by the migration in commit: The
Micropost migration with added index via belongs_to :user, as shown in

app/models/micropost.rb

The second half of the association, has_many :microposts, needs to be
added by hand, as shown in (in the commit).