Testing the micropost order. red
Micropost refinements

In this section, we’ll add a couple of refinements to the user/micropost
association. In particular, we’ll arrange for a user’s microposts to be
retrieved in a specific order, and we’ll also make microposts dependent
on users so that they will be automatically destroyed if their
associated user is destroyed.

Default scope

By default, the user.microposts method makes no guarantees about the
order of the posts, but (following the convention of blogs and Twitter)
we want the microposts to come out in reverse order of when they were
created so that the most recent post is first. We’ll arrange for this to
happen using a default scope.

This is exactly the sort of feature that could easily lead to a spurious
passing test (i.e., a test that would pass even if the application code
were wrong), so we’ll proceed using test-driven development to be sure
we’re testing the right thing. In particular, let’s write a test to
verify that the first micropost in the database is the same as a fixture
micropost we’ll call most_recent, as shown in this commit

```
$ rails test
```

ERROR["test_order_should_be_most_recent_first", MicropostTest,
23.021545933908783]
 test_order_should_be_most_recent_first#MicropostTest (23.02s)
StandardError:         StandardError: No fixture named 'most_recent'
found for fixture set 'microposts'
            test/models/micropost_test.rb:31:in `block in
<class:MicropostTest>'

===E