Micropost fixtures.
In commit: Testing the micropost order, relies on having some micropost
fixtures, which we define as shown in this commit.

With the code in commit: Testing the micropost order and commit:
Micropost fixtures, the test suite should be red:

```
rails test test/models/micropost_test.rb
```

Note that we have explicitly set the created_at column using embedded
Ruby. Because it’s a “magic” column automatically updated by Rails,
setting it by hand isn’t ordinarily possible, but it is possible in
fixtures. In practice this might not be necessary, and in fact on many
systems the fixtures are created in order. In this case, the final
fixture in the file is created last (and hence is most recent), but it
would be foolish to rely on this behavior, which is brittle and probably
system-dependent.

```
order(:created_at)
```

Unfortunately, this orders the results in ascending order from smallest
to biggest, which means that the oldest microposts come out first. To
pull them out in reverse order, we can push down one level deeper and
include a string with some raw SQL:

```
order('created_at DESC')
```

Here DESC is SQL for “descending”, i.e., in descending order from newest
to oldest.5 In older versions of Rails, using this raw SQL used to be
the only option to get the desired behavior, but as of Rails 4.0 we can
use a more natural pure-Ruby syntax as well:

```
order(created_at: :desc)
```

Adding this in a default scope for the Micropost model gives