Ordering the microposts with default_scope. green
 In this commit,  introduces the “stabby lambda” syntax for an object
called a Proc (procedure) or lambda, which is an anonymous function (a
function created without a name). The stabby lambda -> takes in a block
(Section 4.3.2: Blocks) and returns a Proc, which can then be evaluated
with the call method. We can see how it works at the console:

```
>> -> { puts "foo" }
=> #<Proc:0x007fab938d0108@(irb):1 (lambda)>
>> -> { puts "foo" }.call
foo
=> nil
```

 (This is a somewhat advanced Ruby topic, so don’t worry if it doesn’t
make sense right away.)

With the code in this code, the tests should be green:

```
$ rails test
```