A partial for showing a single micropost.
Although we don’t yet have a way to create microposts through the
web—that comes in Section 13.3.2: Creating microposts—this won’t stop us
from displaying them (and testing that display). Following Twitter’s
lead, we’ll plan to display a user’s microposts not on a separate
microposts index page but rather directly on the user show page itself,
as mocked up in Figure 13.4. We’ll start with fairly simple ERb
templates for adding a micropost display to the user profile, and then
we’ll add microposts to the seed data from Section 10.3.2: Sample users
so that we have something to display.

Rendering microposts

Our plan is to display the microposts for each user on their respective
profile page (show.html.erb), together with a running count of how many
microposts they’ve made. As we’ll see, many of the ideas are similar to
our work in Section 10.3: Showing all users on showing all users.

In case you’ve added some microposts in the exercises, it’s a good idea
to reset and reseed the database at this time:

```
$ rails db:migrate:reset
$ rails db:seed
```

Although we won’t need the Microposts controller until Section 13.3:
Manipulating microposts, we will need the views directory in just a
moment, so let’s generate the controller now:

```
$ rails generate controller Microposts
```

Our primary purpose in this section is to render all the microposts for
each user. We saw in Section 10.3.5: Partial refactoring that the code

```
<ul class="users">
  <%= render @users %>
</ul>
```

automatically renders each of the users in the @users variable using the
_user.html.erb partial. We’ll define an analogous _micropost.html.erb
partial so that we can use the same technique on a collection of
microposts as follows:

```
<ol class="microposts">
  <%= render @microposts %>
</ol>
```

Note that we’ve used the ordered list tag ol (as opposed to an unordered
list ul) because microposts are listed in a particular order
(reverse-chronological). The corresponding partial appears in this
commit.

This uses the awesome time_ago_in_words helper method, whose meaning is
probably clear and whose effect we will see in Section 13.2.2: Sample
microposts. In this commit also adds a CSS id for each micropost using

```
<li id="micropost-<%= micropost.id %>">
```

This is a generally good practice, as it opens up the possibility of
manipulating individual microposts at a future date (using JavaScript,
for example).