Adding an @microposts instance variable to the user show action.
The next step is to address the difficulty of displaying a potentially
large number of microposts. We’ll solve this problem the same way we
solved it for users in Section 10.3.3: Pagination, namely, using
pagination. As before, we’ll use the will_paginate method:

```
<%= will_paginate @microposts %>
```

If you compare this with the analogous line on the user index page, in
commit: The users index with pagination, you’ll see that before we had
just

```
<%= will_paginate %>
```

This worked because, in the context of the Users controller,
will_paginate assumes the existence of an instance variable called
@users (which, as we saw in Section 10.3.3: Pagination, should be of
class ActiveRecord::Relation). In the present case, since we are still
in the Users controller but want to paginate microposts instead, we’ll
pass an explicit @microposts variable to will_paginate. Of course, this
means that we will have to define such a variable in the user show
action (in this commit).

Notice here how clever paginate is—it even works through the microposts
association, reaching into the microposts table and pulling out the
desired page of microposts.

Our final task is to display the number of microposts for each user,
which we can do with the count method:

```
user.microposts.count
```

As with paginate, we can use the count method through the association.
In particular, count does not pull all the microposts out of the
database and then call length on the resulting array, as this would
become inefficient as the number of microposts grew. Instead, it
performs the calculation directly in the database, asking the database
to count the microposts with the given user_id (an operation for which
all databases are highly optimized). (In the unlikely event that finding
the count is still a bottleneck in your application, you can make it
even faster using a counter cache.)