Adding microposts to the sample data
With all the work making templates for user microposts in Section
13.2.1: Rendering microposts, the ending was rather anticlimactic. We
can rectify this sad situation by adding microposts to the seed data
from Section 10.3.2: Showing all users.

Adding sample microposts for all the users actually takes a rather long
time, so first we’ll select just the first six users (i.e., the five
users with custom Gravatars, and one with the default Gravatar) using
the take method:

```
User.order(:created_at).take
```

The call to order ensures that we find the first six users that were
created.

For each of the selected users, we’ll make 50 microposts (plenty to
overflow the pagination limit of 30). To generate sample content for
each micropost, we’ll use the Faker gem’s handy Lorem.sentence method.
The result is the new seed data method shown in this commit. (The reason
for the order of the loops in this commit is to intermix the microposts
for use in the status feed (Section 14.3: The status feed). Looping over
the users first gives feeds with big runs of microposts from the same
user, which is visually unappealing.)

At this point, we can reseed the development database as usual:

```
$ rails db:migrate:reset
$ rails db:seed
```

You should also quit and restart the Rails development server.

With that, we are in a position to enjoy the fruits of our Section
13.2.1: Rendering microposts labors by displaying information for each
micropost.