Micropost fixtures with user associations.
Because newly activated users get redirected to their profile pages, we
already have a test that the profile page renders correctly (commit:
Adding account activation to the user signup test.). In this section,
we’ll write a short integration test for some of the other elements on
the profile page, including the work from this section. We’ll start by
generating an integration test for the profiles of our site’s users:

```
$ rails generate integration_test users_profile
      invoke  test_unit
      create    test/integration/users_profile_test.rb
```

To test the micropost display on the profile, we need to associate the
fixture microposts with a user. Rails includes a convenient way to build
associations in fixtures, like this:

```
orange:
  content: "I just ate an orange!"
  created_at: <%= 10.minutes.ago %>
  user: michael
```

By identifying the user as michael, we tell Rails to associate this
micropost with the corresponding user in the users fixture:

```
michael:
  name: Michael Example
  email: michael@example.com
  .
  .
  .
```
To test micropost pagination, we’ll also generate some additional
micropost fixtures using the same embedded Ruby technique we used to
make additional users in commit: Adding 30 extra users to the fixture:

```
<% 30.times do |n| %>
micropost_<%= n %>:
  content: <%= Faker::Lorem.sentence(5) %>
  created_at: <%= 42.days.ago %>
  user: michael
<% end %>
```

Putting all this together gives the updated micropost fixtures in this
commit.