Routes for the Microposts resource.
Having finished both the data modeling and display templates for
microposts, we now turn our attention to the interface for creating them
through the web. In this section, we’ll also see the first hint of a
status feed—a notion brought to full fruition in Chapter 14:mFollowing
users. Finally, as with users, we’ll make it possible to destroy
microposts through the web.

There is one break with past convention worth noting: the interface to
the Microposts resource will run principally through the Profile and
Home pages, so we won’t need actions like new or edit in the Microposts
controller; we’ll need only create and destroy. This leads to the routes
for the Microposts resource shown in this commit. The code in this
commit leads in turn to the RESTful routes shown in Table 13.2, which is
a small subset of the full set of routes seen in Table 2.3. Of course,
this simplicity is a sign of being more advanced, not less—we’ve come a
long way since our reliance on scaffolding in Chapter 2: A toy app, and
we no longer need most of its complexity.