Moving the logged_in_user method into the Application controller. red
Writing the application code needed to get the tests in commit:
Authorization tests for the Microposts controller to pass requires a
little refactoring first. Recall from Section 10.2.1: Requiring
logged-in users that we enforced the login requirement using a before
filter that called the logged_in_user method (commit:  Adding a
logged_in_user before filter). At the time, we needed that method only
in the Users controller, but now we find that we need it in the
Microposts controller as well, so we’ll move it into the Application
controller, which is the base class of all controllers (Section 4.4.4:
Adding a logged_in_user before filter). The result appears in this
commit.