Adding authorization to the Microposts controller actions. green
With the code in commit: Moving the logged_in_user method into the Application controller, the logged_in_user method is now available in the Microposts controller, which means that we can add create and destroy actions and then restrict access to them using a before filter, as shown in this commit.

At this point, the tests should pass:

```
$ rails test
```

Exercises -

Why is it a bad idea to leave a copy of logged_in_user in the Users controller?

Its a bad idea to leave a copy of logged_in_user in the Users controller because we may up the application controller version and not update User controller version that leads to be out sync code plus its DRY.