The Microposts controller create action.
In Chapter 7: Sign up, we implemented user signup by making an HTML form that issued an HTTP POST request to the create action in the Users controller. The implementation of micropost creation is similar; the main difference is that, rather than using a separate page at /microposts/new, we will put the form on the Home page itself (i.e., the root path /), as mocked up in Figure 13.10.

When we last left the Home page, it appeared as in the browser (http://localhost:3000/) — that is, it had a “Sign up now!” button in the middle. Since a micropost creation form makes sense only in the context of a particular logged-in user, one goal of this section will be to serve different versions of the Home page depending on a visitor’s login status. We’ll implement this in commit: Adding microposts creation to the Home page.

We’ll start with the create action for microposts, which is similar to its user analogue (commit: The user create action with a save and a redirect); the principal difference lies in using the user/micropost association to build the new micropost, as seen in this commit. Note the use of strong parameters via micropost_params, which permits only the micropost’s content attribute to be modified through the web.