The partial for the user info sidebar.
Note that, as in the profile sidebar (commit: Adding microposts to the
user show page), the user info in commit: The partial for the user info
sidebar displays the total number of microposts for the user. There’s a
slight difference in the display, though; in the profile sidebar,
“Microposts” is a label, and showing “Microposts (1)” makes sense. In
the present case, though, saying “1 microposts” is ungrammatical, so we
arrange to display “1 micropost” and “2 microposts” using the pluralize
method we saw in Section 7.3.3: Signup error messages.

We next define the form for creating microposts (in this commit), which
is similar to the signup form in commit: A form to sign up new users.