Error messages that work with other objects. red
The second change needed to get commit: The form partial for creating
microposts to work is to redefine the error-messages partial so the
following code from commit: The form partial for creating microposts
works:

```
<%= render 'shared/error_messages', object: f.object %>
```

You may recall from commit: Code to display error messages on the signup
form that the error-messages partial references the @user variable
explicitly, but in the present case we have an @micropost variable
instead. To unify these cases, we can pass the form variable f to the
partial and access the associated object through f.object, so that in

```
form_for(@user) do |f|
```

f.object is @user, and in

```
form_for(@micropost) do |f|
```
f.object is @micropost, etc.

To pass the object to the partial, we use a hash with value equal to the
object and key equal to the desired name of the variable in the partial,
which is what the second line in commit: The form partial for creating
microposts accomplishes. In other words, object: f.object creates a
variable called object in the error_messages partial, and we can use it
to construct a customized error message, as shown in this commit.

At this point, you should verify that the test suite is red:

```
$ rails test
```

ERROR["test_password_resets", PasswordResetsTest, 25.74674785195384]
 test_password_resets#PasswordResetsTest (25.75s)
ActionView::Template::Error:         ActionView::Template::Error:
undefined local variable or method `object' for
#<#<Class:0x0000000d933f48>:0x0000000e0705e0>
            app/views/shared/_error_messages.html.erb:1:in
`_app_views_shared__error_messages_html_erb___1467367106_116627780'
            app/views/password_resets/edit.html.erb:7:in `block in
_app_views_password_resets_edit_html_erb__1030498844_117672540'
            app/views/password_resets/edit.html.erb:6:in
`_app_views_password_resets_edit_html_erb__1030498844_117672540'
            test/integration/password_resets_test.rb:38:in `block in
<class:PasswordResetsTest>'

=E