Updating the rendering of user signup errors.
This is a hint that we need to update the other occurrences of the
error-messages partial, which we used when signing up users (commit:
Code to display error messages on the signup form), resetting passwords
(commit: The form to reset a password), and editing users (commit: The
user edit view). The updated versions are shown in Listing 13.43,
Listing 13.45, and Listing 13.44.

Listing 13.43: Updating the rendering of user signup errors.
app/views/users/new.html.erb

Listing 13.44: Updating the errors for editing users.
app/views/users/edit.html.erb

Listing 13.45: Updating the errors for password resets.
app/views/password_resets/edit.html.erb

Because we've created a partial that is included in Listing 13.43 and
Listing 13.45; above so there isn't a lot re-coding need.

Good example of DRY design.

At this point, all the tests should be green:

```
$ rails test
```

Exercises -

1. Refactor the Home page to use separate partials for the two branches
of the if-else statement.

Created two new partials:  login_user and new_user then rendered them in
the if-then-else block as so:

```
<% if logged_in? %>
  <%= render 'login_user' %>
<% else %>
  <%= render 'new_user' %>
<% end %>
```